% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_w_mean}
\alias{array_w_mean}
\title{Index based weighted means from arrays in C++}
\usage{
array_w_mean(array, scheme)
}
\arguments{
\item{array}{to be summarised.}

\item{scheme}{a dataframe containing the indices to average together with weights.}
}
\value{
A vector containing summary values per group, ordered by group.
}
\description{
This function takes an array and a summary scheme and returns a vector of weighted means per group.
}
\details{
A summary scheme must contain 3 columns of array indices on rows, columns and slices, titled "x", "y", 
and "layer". A column named "group" defines the values to be summarised together. A column "weight" 
contains the weights for computing weighted mean. Helper functions are coming to help make common schemes.
}
