% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{calculate_proximity_weight}
\alias{calculate_proximity_weight}
\title{Calculate the Weights for Linear Interpolation Between Two Depths}
\usage{
calculate_proximity_weight(depths, target)
}
\arguments{
\item{depths}{A numeric vector of a shallower and deeper depth.}

\item{target}{An intermediate depth we would like to interpolate to.}
}
\value{
A vector of weights to match the vector of depths.
}
\description{
This function calculates the distance between a shallower and deeper depth to a target depth. These values 
are then swapped so they can be used in a weighted average to linearly interpolate to the target depth.
}
\examples{
calculate_proximity_weight(c(0, 100), target = 30)

calculate_proximity_weight(c(0, 100), target = 50)

calculate_proximity_weight(c(0, 100), target = 80)
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{calculate_depth_share}()},
\code{\link{netcdf_scheme_helpers}},
\code{\link{scheme_column}()},
\code{\link{scheme_interp_slice}()},
\code{\link{scheme_strathE2E}()},
\code{\link{stratify}()},
\code{\link{voronoi_grid}()},
\code{\link{xyindex_to_nindex}()}
}
\concept{NEMO-MEDUSA spatial tools}
