% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary functions.R
\name{check_grid}
\alias{check_grid}
\title{Check if a segment runs parallel to latitude or longitude}
\usage{
check_grid(segment, Weighted)
}
\arguments{
\item{segment}{An SF line object representing a transect along the perimeter of a model domain polygon.}
}
\value{
The function returns a dataframe detailing whether laitudues are the same and longitudes are the same along a segment.
}
\description{
This function checks whether the transects created to sample the boundaries of model compartments lie on
a lat-lon grid. If they do, then we know that when sampling meridional and zonal currents only one is needed
per transect.
}
\details{
The function takes a boundary segment (transect) and reprojects it on a lat-lon grid. The coordinates are extracted,
and checked for identical latitudes or longitudes.
}
\seealso{
Other Boundary sampling functions: 
\code{\link{Sample_OOB}()},
\code{\link{Sample}()},
\code{\link{boundaries}()},
\code{\link{direction}()},
\code{\link{extract_OOB}()},
\code{\link{extract}()},
\code{\link{to_segments}()}
}
\concept{Boundary sampling functions}
