% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{detritus_month}
\alias{detritus_month}
\title{Condense Daily netcdf Files into a Monthly Summary}
\usage{
detritus_month(data, targets, ...)
}
\arguments{
\item{data}{A dataframe containing the metadata of multiple netcdf files from a common month.}

\item{targets}{A dataframe containing the points from the grid we want to retain data for.}

\item{...}{Additional arguments passed on to `get_detritus`.}
}
\value{
The function returns a dataframe containing the monthly average shalllow and deep spatial grids for
detrital nitrogen.
}
\description{
This function is a variant of `whole_month` which only considers detritus files. This is a convenience function.
Detritus data is available on it's own grid, so there's no reason to pass to `type_in_month` and specify the
required metadata.
}
\details{
The function takes the metadata for a collection of files which contain data from the same month. The files
are passed to `get_detritus` to be summarised, reducing the NEMO-MEDUSA model outputs from large arrays to effectively
two matrices for the month. Points outside the project window are removed before saving the dataframe for the month in
"./Objects/Detritus/".

Creating intermediate monthly objects allows the terabytes of data to be reduced without encountering memory issues.
Also, working on independent monthly packets of data means we can parallelise any data processing for speed.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_bio}()},
\code{\link{get_detritus}()},
\code{\link{get_ice}()},
\code{\link{get_merid}()},
\code{\link{get_rivers}()},
\code{\link{get_sea}()},
\code{\link{get_spatial}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{interp_month}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
