% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary functions.R
\name{direction}
\alias{direction}
\title{Test whether "positive" currents flow in or out of a model compartment at a segment}
\usage{
direction(segment)
}
\arguments{
\item{segment}{An SF line object representing a transect along the perimeter of a model domain polygon.}
}
\value{
The function returns a dataframe detailing which two model compartments are either side of the transect,
and the direction of the exchange.
}
\description{
This function takes a transect from the edge of a model compartment and works out where it is spatially. This is
neccessary for correctly averaging variables along the whole boundary of a compartment.
}
\details{
The function determines which model compartment is next to the focal compartment. The focal compartment is either the
offshore or inshore polygon, with the offshore polygon selected ahead of the inshore polygon when these share a boundary.
This defines the link between model compartments.

Direction (in/out) is also calculated relative to the focal polygon. At the western boundary of a polygon, positive zonal
currents \emph{(West to East)} indicate water flowing into the model compartment. However, at the Eastern boundary,
flows from West to East now leave the model compartment. A True or False column called flip is included. If a positive value
for a water velocity is leaving the polygon, or a negative flow value is entering the polygon, flip can be used to multiply
by -1 to correctly sum the water movements from the perspective of the focal polygon.
}
\seealso{
Other Boundary sampling functions: 
\code{\link{Sample_OOB}()},
\code{\link{Sample}()},
\code{\link{boundaries}()},
\code{\link{check_grid}()},
\code{\link{extract_OOB}()},
\code{\link{extract}()},
\code{\link{to_segments}()}
}
\concept{Boundary sampling functions}
