% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions 1D.R
\name{extractors_1D}
\alias{extractors_1D}
\alias{get_grid_T_1D}
\alias{get_ptrc_T_1D}
\alias{get_grid_W_1D}
\alias{get_grid_V_1D}
\alias{get_grid_U_1D}
\title{Extract a 1D water column from NEMO-MEDUSA}
\usage{
get_grid_T_1D(filename, date, start, count)

get_ptrc_T_1D(filename, date, x, y)

get_grid_W_1D(filename, date, x, y)

get_grid_V_1D(filename, date, x, y)

get_grid_U_1D(filename, date, x, y)
}
\arguments{
\item{filename}{The name of a netcdf file containing the title variables.}

\item{date}{The numeric string representing the date contained by the data file.}

\item{start}{an optional vector of indices to start subsetting at. See ncdf4 documentation.}

\item{count}{an optional vector of steps to subset along. See ncdf4 documentation.}

\item{...}{soaks up unused function arguments passed by the wrapper functions handling file architecture.}
}
\value{
A dataframe containing estimates of all variables at all depths for a single point in the model grid
}
\description{
These functions read in target variables from NEMO-MEDUSA model outputs and return a water column.
}
\details{
Each variable of interest in the netcdf file is imported, only reading at a single specified pixel for all depths.

Each model output file type contains a different set of variables, extracted by the relevant function variant:\tabular{ll}{
   File type         \tab Variables \cr
   grid_T_ \tab Salinity, temperature, sea ice concentration. \cr
   grid_U_ \tab Zonal currents. \cr
   grid_V_ \tab Meridional currents. \cr
   grid_W_ \tab Vertical velocity, vertical eddy diffusivitiy. \cr
   ptrc_T_ \tab DIN, phytoplankton nitrogen content, phytoplankton chlorophyll, <br> zooplankton nitrogen content, detritus, silicate. \cr
}


Some function variants have different arguments:

grid_W_ expects it's own \code{scheme_w} as depth levels are different between these and other files.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{NEMO_MEDUSA}()},
\code{\link{extractors_slabR}},
\code{\link{get_grid_T_StrathE2E}()},
\code{\link{get_grid_U_StrathE2E}()},
\code{\link{get_grid_V_StrathE2E}()},
\code{\link{get_grid_W_StrathE2E}()},
\code{\link{get_icemod_StrathE2E}()},
\code{\link{get_ptrc_T_StrathE2E}()},
\code{\link{get_spatial}()}
}
\concept{NEMO-MEDUSA variable extractors}
