% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions STRATHE2E.R
\name{extractors_SE2E}
\alias{extractors_SE2E}
\alias{get_grid_T_StrathE2E}
\alias{get_ptrc_T_StrathE2E}
\alias{get_icemod_StrathE2E}
\alias{get_grid_W_StrathE2E}
\alias{get_grid_V_StrathE2E}
\alias{get_grid_U_StrathE2E}
\title{DEPRECATED. Extract a StrathE2E summary from NEMO-MEDUSA 

These functions read in target variables from NEMO-MEDUSA model outputs and return a summarised slab.}
\usage{
get_grid_T_StrathE2E(path, file, space)

get_ptrc_T_StrathE2E(path, file, space)

get_icemod_StrathE2E(path, file, space)

get_grid_W_StrathE2E(path, file, space)

get_grid_V_StrathE2E(path, file, space)

get_grid_U_StrathE2E(path, file, space)
}
\description{
DEPRECATED. Extract a StrathE2E summary from NEMO-MEDUSA 

These functions read in target variables from NEMO-MEDUSA model outputs and return a summarised slab.
}
