% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{get_air}
\alias{get_air}
\title{Get Surface Irradiance & Air Temperature}
\usage{
get_air(File, Type, Year)
}
\arguments{
\item{File}{The location of the netcdf file containing NEMO-MEDUSA driving data.}

\item{Type}{The variable contained in the netcdf file either "T150" (air temperature) or "SWF" (surface irradiance).}

\item{Year}{The year the necdf file contains data for.}
}
\value{
A dataframe containing a monthly time series within a year of either average air temperature or surface
irradiance. Air temperature is also split by shore zone.
}
\description{
This function reads either title variable from a NEMO-MEDUSA model \strong{DRIVERS} file, and returns a monthly time series.
}
\details{
The appropriate variable in the netcdf file is imported according to the `Type` parameter, only reading within an
x/y window specified in `Space`. The function then drops points outside the model domain before constructing the monthly
time series.

Unlike other NEMO-MEDUSA related get_* functions, this function constructs the monthly time series directly. A wrapper function
to handle time isn't required, as each netcdf file for driving data contains 360 day steps for a single year. `stratify` also
isn't called, as these variables have no depth dimension.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{detritus_month}()},
\code{\link{get_air_dt}()},
\code{\link{get_bio}()},
\code{\link{get_detritus}()},
\code{\link{get_ice}()},
\code{\link{get_merid}()},
\code{\link{get_rivers}()},
\code{\link{get_sea}()},
\code{\link{get_spatial}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{interp_month}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
