% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{get_detritus}
\alias{get_detritus}
\title{Get Detritus}
\usage{
get_detritus(file, start, count, grid, shallow, s.weights, deep, d.weights)
}
\arguments{
\item{file}{The name of a netcdf file containing the title variables.}

\item{start}{The indices to start reading data from in 3 dimensions.}

\item{count}{The lengths of data reads in 3 dimensions.}

\item{grid}{A dataframe to bind the extracted values to.}

\item{shallow, }{A TRUE / FALSE vector indicating which depth layers should be combined into a shallow zone.}

\item{s.weights}{An array of water thicknesses to use as weights for collapsing layers into a shallow zone.}

\item{deep, }{A TRUE / FALSE vector indicating which depth layers should be combined into a deep zone.}

\item{d.weights}{An array of water thicknesses to use as weights for collapsing layers into a deep zone.}
}
\value{
A dataframe containing points and lat/lon coordinates, with the average NEMO-MEDUSA model outputs for the title variable
in the shallow and deep zone. The dataframe contains the data for a single day.
}
\description{
This function reads in the title variable from NEMO-MEDUSA model outputs and reshapes for StrathE2E.
}
\details{
The variable of interest in the netcdf file is imported, only reading within an x/y window specified with `start3D` and
`count3D`. The values are then passed to `stratify` to calculate two average matrices, one a weighted vertical average
of the shallow zone, and the same for the deep zone. Latitude and longitudes are also attached to each horizontal pixel.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{detritus_month}()},
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_bio}()},
\code{\link{get_ice}()},
\code{\link{get_merid}()},
\code{\link{get_rivers}()},
\code{\link{get_sea}()},
\code{\link{get_spatial}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{interp_month}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
