% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions.R
\name{get_grid_U_}
\alias{get_grid_U_}
\title{Get Zonal Currents}
\usage{
get_grid_U_(path, file, space)
}
\arguments{
\item{path}{The path to the NEMO-MEDUSA model outputs.}

\item{file}{The name of a netcdf file containing the title variables.}
}
\value{
A dataframe containing points and lat/lon coordinates, with the average NEMO-MEDUSA model outputs for the title variable
in the shallow and deep zone. The dataframe contains the data for a single day.
}
\description{
This function reads in the title variable from NEMO-MEDUSA model outputs and reshapes for StrathE2E.
Zonal currents can be found in files labelled "grid_U_".
}
\details{
The variable of interest in the netcdf file is imported, only reading within an x/y window specified with `start3D` and
`count3D`. The values are then passed to `stratify` to calculate two average matrices, one a weighted vertical average
of the shallow zone, and the same for the deep zone. Latitude and longitudes are also attached to each horizontal pixel.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_grid_T_}()},
\code{\link{get_grid_V_}()},
\code{\link{get_grid_W_}()},
\code{\link{get_icemod_}()},
\code{\link{get_ptrc_T_}()},
\code{\link{get_rivers}()},
\code{\link{get_spatial}()},
\code{\link{interp_month}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
