% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions STRATHE2E.R
\name{get_grid_W_StrathE2E}
\alias{get_grid_W_StrathE2E}
\title{Get Vertical Velocity and Vertical Eddy Diffusivity}
\usage{
get_grid_W_StrathE2E(path, file, space)
}
\arguments{
\item{path}{The path to the NEMO-MEDUSA model outputs.}

\item{file}{The name of a netcdf file containing the title variables.}
}
\value{
A dataframe containing points and lat/lon coordinates, with the average NEMO-MEDUSA model outputs for title variables
in the shallow and deep zone. The dataframe contains the data for a single day.
}
\description{
This function reads in the title variables from NEMO-MEDUSA model outputs and reshapes for StrathE2E.
Vertical velocity and vertical eddy diffusivitiy can be found in files labelled "W".
}
\details{
Each variable of interest in the netcdf file is imported, only reading within an x/y window specified with `start3D` and
`count3D`. The values are then passed to `stratify` to calculate two average matrices, one a weighted vertical average
of the shallow zone, and the same for the deep zone. Latitude and longitudes are also attached to each horizontal pixel.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{NEMO_MEDUSA}()},
\code{\link{extractors_1D}},
\code{\link{extractors_slabR}},
\code{\link{get_grid_T_StrathE2E}()},
\code{\link{get_grid_U_StrathE2E}()},
\code{\link{get_grid_V_StrathE2E}()},
\code{\link{get_icemod_StrathE2E}()},
\code{\link{get_ptrc_T_StrathE2E}()},
\code{\link{get_spatial}()}
}
\concept{NEMO-MEDUSA variable extractors}
