% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions 1D.R
\name{get_ptrc_T_1D}
\alias{get_ptrc_T_1D}
\title{Get Dissolved Inorganic Nitrogen & Chlorophyll a}
\usage{
get_ptrc_T_1D(filename, date, x, y)
}
\arguments{
\item{filename}{The name of a netcdf file containing the title variables.}

\item{date}{The numeric string representing the date contained by the data file.}

\item{x}{The x index in the grid to sample at.}

\item{y}{The y index in the grid to sample at.}
}
\value{
A dataframe containing estimates of all variables at all depths for a single point in the model grid
}
\description{
This function reads in the title variables from NEMO-MEDUSA model outputs and returns a water column.
}
\details{
Each variable of interest in the netcdf file is imported, only reading at a single specified pixel for all depths.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_grid_T_1D}()},
\code{\link{get_grid_T_StrathE2E}()},
\code{\link{get_grid_U_1D}()},
\code{\link{get_grid_U_StrathE2E}()},
\code{\link{get_grid_V_1D}()},
\code{\link{get_grid_V_StrathE2E}()},
\code{\link{get_grid_W_1D}()},
\code{\link{get_grid_W_StrathE2E}()},
\code{\link{get_icemod_StrathE2E}()},
\code{\link{get_ptrc_T_StrathE2E}()},
\code{\link{get_rivers}()},
\code{\link{get_spatial}()},
\code{\link{interp_month}()},
\code{\link{type_in_month_old}()},
\code{\link{type_in_month}()},
\code{\link{whole_month_old}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
