% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions slabR.R
\name{get_ptrc_T_slabR}
\alias{get_ptrc_T_slabR}
\title{Get Dissolved Inorganic Nitrogen & Phytoplankton Nitrogen}
\usage{
get_ptrc_T_slabR(path, file, scheme, space, ...)
}
\arguments{
\item{path}{The path to the NEMO-MEDUSA model outputs.}

\item{file}{The name of a netcdf file containing the title variables.}
}
\value{
A dataframe containing points and lat/lon coordinates, with the average NEMO-MEDUSA model outputs for title variables
in the shallow and deep zone. The dataframe contains the data for a single day.
}
\description{
This function reads in the title variables from NEMO-MEDUSA model outputs and reshapes for StrathE2E.
DIN & phytoplankton Nitrogen content can be found in files labelled "ptrc_T_".
}
\details{
Each variable of interest in the netcdf file is imported, only reading within an x/y window specified with `start3D` and
`count3D`. The values are then passed to `array_w_mean` to calculate two average matrices, one a weighted vertical average
of the shallow zone, and the same for the deep zone. Latitude and longitudes are also attached to each horizontal pixel.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_grid_T_1D}()},
\code{\link{get_grid_T_StrathE2E}()},
\code{\link{get_grid_T_slabR}()},
\code{\link{get_grid_U_1D}()},
\code{\link{get_grid_U_StrathE2E}()},
\code{\link{get_grid_U_slabR}()},
\code{\link{get_grid_V_1D}()},
\code{\link{get_grid_V_StrathE2E}()},
\code{\link{get_grid_V_slabR}()},
\code{\link{get_grid_W_1D}()},
\code{\link{get_grid_W_StrathE2E}()},
\code{\link{get_grid_W_slabR}()},
\code{\link{get_icemod_StrathE2E}()},
\code{\link{get_icemod_slabR}()},
\code{\link{get_ptrc_T_1D}()},
\code{\link{get_ptrc_T_StrathE2E}()},
\code{\link{get_rivers}()},
\code{\link{get_spatial}()},
\code{\link{interp_month}()},
\code{\link{type_in_month_old}()},
\code{\link{type_in_month}()},
\code{\link{whole_month_old}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
