% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting functions.R
\name{jacks_polar_animate}
\alias{jacks_polar_animate}
\title{Animate a map of GFW data for a geartype, above a minimum level of fishing activity (polar projection)}
\usage{
jacks_polar_animate(gear, limit)
}
\arguments{
\item{gear}{The gear type used by boats to be mapped.}

\item{limit}{The minimum number of hours of fishing activity required for a pixel to be included in the map.}
}
\value{
The function saves a map of fishing activity for a particular gear type above a minimum level of activity.
}
\description{
This functions saves an animated map of fishing activity by month. with a CRS appropriate for the arctic.
}
\details{
The function saves a map of fishing activity, as recorded by global fishing watch, with FAO fishing regions overlayed.
The user can specify which fishing gear type to visualise, and can also remove points below a certain amount
of fishing effort. This is useful for keeping the plot clean, but also allows you to remove any fishing signals
you might not trust from the global fishing watch model GFW model.
}
\seealso{
Other Global Fishing Watch functions: 
\code{\link{jacks_lazy_animation}()},
\code{\link{jacks_lazy_avg}()},
\code{\link{jacks_polar_static}()}
}
\concept{Global Fishing Watch functions}
