% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemomedusR-package.R
\docType{package}
\name{nemomedusR-package}
\alias{nemomedusR}
\alias{nemomedusR-package}
\title{nemomedusR: Tools for summarising NEMO-MEDUSA model output}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package helps produce common spatial and temporal summaries from NEMO-MEDUSA model output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jack-h-laverick.github.io/nemomedusR/}
  \item \url{https:github.com/Jack-H-Laverick/nemomedusR}
  \item Report bugs at \url{https:github.com/Jack-H-Laverick/nemomedusR/issues}
}

}
\author{
\strong{Maintainer}: Jack H Laverick \email{jacklaverick@ymail.com} (\href{https://orcid.org/0000-0001-8829-2084}{ORCID})

}
\keyword{internal}
