% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting functions.R
\name{point_plot}
\alias{point_plot}
\title{Map Spatial Summaries}
\usage{
point_plot(data, var, zoom)
}
\arguments{
\item{data}{A dataframe containing the columns "Zonal" and "Meridional" for currents.}

\item{var}{A "quoted" column name denoting how to colour the map.}

\item{zoom}{A call to `coord_sf` specifying the project plotting window.}
}
\value{
The function creates a facetted map of a summarised variable. Each facet shows a month.
The plot is saved into the NEMO-MEDUSA/grids folder.
}
\description{
This function builds a map of the passed variable, facetted by month.
}
\seealso{
Other NEMO-MEDUSA plots: 
\code{\link{point_plot_voronoi}()},
\code{\link{stick_plot}()},
\code{\link{ts_plot}()}
}
\concept{NEMO-MEDUSA plots}
