% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary functions.R
\name{Sample}
\alias{Sample}
\title{Sample water flows across transects along the boundaries of model compartments}
\usage{
Sample(file, ...)
}
\arguments{
\item{file}{Path to the .rds object containing both Zonal and Meridional water velocities.}

\item{...}{Additional arguments to pass to `extract`.}
}
\value{
The function returns a dataframe of net water movements between model compartments
and the external environment in a given month.
}
\description{
This function calculates the water exchanges between model compartments and external boundaries.
}
\details{
A datafile containing water velocities is read in and split by depth before calling `extract` to
pull the correct velocities.

`Sample` then groups exchanges by depth, shore zone, neighbour and direction. Flows
are summed to return net water movements for the given time step.
}
\seealso{
Other Boundary sampling functions: 
\code{\link{Sample_OOB}()},
\code{\link{boundaries}()},
\code{\link{check_grid}()},
\code{\link{direction}()},
\code{\link{extract_OOB}()},
\code{\link{extract}()},
\code{\link{to_segments}()}
}
\concept{Boundary sampling functions}
