% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schemes.R
\name{scheme_column}
\alias{scheme_column}
\title{Return a slabR Scheme for a StrathE2E Summary}
\usage{
scheme_column(space, bathymetry, sf = NULL)
}
\arguments{
\item{space}{a list of latitudes, longitudes, and depths for NEMO-MEDUSA model output, as returned by `get_spatial()`}

\item{bathymetry}{a dataframe containg the NEMO_MEDUSA bathymetry.}

\item{sf}{an sf point object defining the location to extract a representive water column for.}
}
\value{
A dataframe containing a slabR scheme.
}
\description{
Use this scheme to extract the values for the nearest depth column to a point location.
}
\details{
The function takes the spatial information for a whole NEMO-MEDUSA file and identifies the grid point closest
to a point location (the sf object). The returned sampling scheme contains every depth layer above the seafloor
with an equal weight. To retrieve the average depth column over a time period, use the "slabR" analysis with 
`NEMO_MEDUSA`. To Simply subset the depth column from model output, returning all values, use the "1D" analysis 
option.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{calculate_depth_share}()},
\code{\link{calculate_proximity_weight}()},
\code{\link{netcdf_scheme_helpers}},
\code{\link{scheme_interp_slice}()},
\code{\link{scheme_strathE2E}()},
\code{\link{stratify}()},
\code{\link{voronoi_grid}()},
\code{\link{xyindex_to_nindex}()}
}
\concept{NEMO-MEDUSA spatial tools}
