% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schemes.R
\name{scheme_interp_slice}
\alias{scheme_interp_slice}
\title{Return a slabR Scheme for Linear Interpolation Between Array Slices}
\usage{
scheme_interp_slice(space, target_depth, sf = NULL)
}
\arguments{
\item{space}{a list of latitudes, longitudes, and depths for NEMO-MEDUSA model output, as returned by `get_spatial()`}

\item{target_depth}{a depth (m) to interpolate a slice for.}

\item{sf}{an optional sf polygon to define the horizontal area of interest.}
}
\value{
A dataframe containing a slabR scheme.
}
\description{
Use this scheme to interpolate a layer at a new depth from an array.
}
\details{
The function takes the spatial information for a whole NEMO-MEDUSA file and determines which slice is immediately
above and below a target depth. These two slices are used to interpolate the new layer. The function calculates 
the weights needed to interpolate between the values using `calculate_proximity_weight`. If an sf polygon is provided,
only points which fall within the polygon will be interpolated.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{calculate_depth_share}()},
\code{\link{calculate_proximity_weight}()},
\code{\link{empty}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights}()},
\code{\link{netcdf_scheme_helpers}},
\code{\link{scheme_column}()},
\code{\link{scheme_strathE2E}()},
\code{\link{stratify}()},
\code{\link{voronoi_grid}()},
\code{\link{xyindex_to_nindex}()}
}
\concept{NEMO-MEDUSA spatial tools}
