% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schemes.R
\name{scheme_strathE2E}
\alias{scheme_strathE2E}
\title{Return a slabR Scheme for a StrathE2E Summary}
\usage{
scheme_strathE2E(space, bathymetry, shallow, deep, sf = NULL)
}
\arguments{
\item{space}{a list of latitudes, longitudes, and depths for NEMO-MEDUSA model output, as returned by `get_spatial()`}

\item{bathymetry}{a dataframe containg the NEMO_MEDUSA bathymetry.}

\item{shallow}{a vector of a min and max depth (m) for the shallow layer to be averaged over.}

\item{deep}{a vector of a min and max depth (m) for the deep layer to be averaged over.}

\item{sf}{an optional sf polygon to define the horizontal area of interest.}
}
\value{
A dataframe containing a slabR scheme.
}
\description{
Use this scheme to create a depth-averaged shallow and deep grid for use with StrathE2E.
}
\details{
The function takes the spatial information for a whole NEMO-MEDUSA file and determines how to get a depth-averaged 
shallow and deep grid. The function uses `calculate_depth_share()` per pixel to determine how much of the water column
is represented by each slice, within either the shallow or deep grid. The depth limits of each layer and the 
bathymetry are used when calculating the weights for each point. If an sf polygon is provided, only points which fall 
within the polygon will be interpolated.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{calculate_depth_share}()},
\code{\link{calculate_proximity_weight}()},
\code{\link{netcdf_scheme_helpers}},
\code{\link{scheme_column}()},
\code{\link{scheme_interp_slice}()},
\code{\link{stratify}()},
\code{\link{voronoi_grid}()},
\code{\link{xyindex_to_nindex}()}
}
\concept{NEMO-MEDUSA spatial tools}
