% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Pull Coordinates from a Simple Feature Geometry Column}
\usage{
sfc_as_cols(x, names = c("x", "y"))
}
\arguments{
\item{data}{An SF (Simple Feature) object.}
}
\value{
The same object, now with two columns containing the coordinates in the geometry column.
}
\description{
This function takes an SF object, and adds two columns containing the coordinates in the geometry column.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{Window}()},
\code{\link{calculate_depth_share}()},
\code{\link{empty}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights}()},
\code{\link{reproj}()},
\code{\link{stratify}()},
\code{\link{vectors_2_direction}()},
\code{\link{voronoi_grid}()},
\code{\link{xyindex_to_nindex}()}
}
\concept{NEMO-MEDUSA spatial tools}
