% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting functions.R
\name{stick_plot}
\alias{stick_plot}
\title{Create Stick Plots to Show Currents}
\usage{
stick_plot(data, zoom, pre)
}
\arguments{
\item{data}{A dataframe containing the columns "Zonal" and "Meridional" for currents.}

\item{zoom}{A call to `coord_sf` specifying the project plotting window.}

\item{pre}{A list of predefined preferences for saving maps in the project (scale, width, height, units, dpi)}
}
\value{
The function creates a map of Water velocities. Colour is used to show speed, sticks are included to show direction.
The plot is saved into the NEMO-MEDUSA/Currents folder.
}
\description{
This function builds a stick plot to map water velocities in the project window.
}
\seealso{
Other NEMO-MEDUSA plots: 
\code{\link{point_plot_voronoi}()},
\code{\link{point_plot}()},
\code{\link{ts_plot}()}
}
\concept{NEMO-MEDUSA plots}
