% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{stratify}
\alias{stratify}
\title{Summarise Across Depths in a NEMO-MEDUSA Array}
\usage{
stratify(data, depth, weights)
}
\arguments{
\item{data}{An array containing estimates of a variable.}

\item{depth}{A TRUE FALSE vector indicating the depth layers to be extracted.}

\item{weights}{An array containing water thicknesses.}
}
\value{
A matrix containing the weighted averages of the variable across the depth window of interest.
}
\description{
This function takes an array of a variable, and an array of water thicknesses to perform a weighted average across depth. The depth
window to be averaged can be specified, so this function can be used to create both shallow and deep layers (or more for that matter).
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{calculate_depth_share}()},
\code{\link{empty}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights}()},
\code{\link{voronoi_grid}()},
\code{\link{xyindex_to_nindex}()}
}
\concept{NEMO-MEDUSA spatial tools}
