% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Averages.R
\name{strip_ice}
\alias{strip_ice}
\title{Strip Snow and Ice Variables at Depth}
\usage{
strip_ice(data, dt)
}
\arguments{
\item{data}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs, at a single depth.}

\item{dt}{Switch for using either data.table or dplyr methods (TRUE/FALSE respectively)}
}
\value{
If the `data` contains shallow data, no action is taken. If `data` contains deep data, columns for variables only
relevant in the shallow zone are dropped.
}
\description{
This function removes the snow and ice columns from a dataframe if depth = "D".
}
\details{
Some variables are only relevant in the shallow zone of StrathE2E polar. There is no sea-ice 60 m under the sea.
This means, when dataframes containing both shallow and deep data are split by depth, empty columns can be introduced.
These empty columns can cause problems in downstream functions, such as plotting by column. This function removes the
empty columns.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{decadal}()},
\code{\link{summarise_sp}()},
\code{\link{summarise_ts}()}
}
\concept{NEMO-MEDUSA averages}
