% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{summarise_ts}
\alias{summarise_ts}
\title{Average into Time Series}
\usage{
summarise_ts(saved)
}
\arguments{
\item{saved}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs.}
}
\value{
A dataframe containing a mean monthly time series of all target variables in NEMO-MEDUSA outputs.
}
\description{
This function averages NEMO-MEDUSA monthly summaries into time series for each model compartment.
}
\details{
The function groups by model compartment (Depth and Shore zone) and time step (Month and Year).
The mean for every target variable is calculated within these groups.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{decadal}()},
\code{\link{strip_ice}()},
\code{\link{summarise_sp}()},
\code{\link{summarise_ts_detritus}()}
}
\concept{NEMO-MEDUSA averages}
