% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{summarise_ts_detritus}
\alias{summarise_ts_detritus}
\title{Average into Detrital Time Series}
\usage{
summarise_ts_detritus(saved)
}
\arguments{
\item{saved}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs.}
}
\value{
A dataframe containing a mean monthly time series of detrital nitrogen in NEMO-MEDUSA outputs.
}
\description{
This function is a variant of `summarise_ts` which only considers detritus files.
}
\details{
The function groups by model compartment (Depth and Shore zone) and time step (Month and Year) before calculating
mean detrital nitrogen within groups.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{decadal}()},
\code{\link{strip_ice}()},
\code{\link{summarise_sp}()},
\code{\link{summarise_ts}()}
}
\concept{NEMO-MEDUSA averages}
