% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary functions.R
\name{to_segments}
\alias{to_segments}
\title{Break an SF linestring into segments at corners}
\usage{
to_segments(line, crs)
}
\arguments{
\item{line}{The geographic perimeter of the inshore and offshore zone as SF object.}

\item{crs}{The coordinate reference system for the project. This is inherited from `boundaries` and should be specified in
the project region file.}
}
\value{
The function returns a list of individual SF lines connecting each corner of the perimeters of model compartments.
}
\description{
This function is wrapped within the boundaries function and is used to convert the perimeter of
model zones into seperate segments between each corner. These segments can then be used independently to sample data.
}
\seealso{
Other Boundary sampling functions: 
\code{\link{Sample_OOB}()},
\code{\link{Sample}()},
\code{\link{boundaries}()},
\code{\link{check_grid}()},
\code{\link{direction}()},
\code{\link{extract_OOB}()},
\code{\link{extract}()}
}
\concept{Boundary sampling functions}
