% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting functions.R
\name{ts_plot}
\alias{ts_plot}
\title{Plot Temporal Summaries}
\usage{
ts_plot(var)
}
\arguments{
\item{var}{A "quoted" column name denoting how to colour the map.}
}
\value{
The function creates a time series by model compartment of a summarised variable.
The plot is saved into the NEMO-MEDUSA folder.
}
\description{
This function builds a time series, split by model compartment.
}
\seealso{
Other NEMO-MEDUSA plots: 
\code{\link{point_plot_voronoi}()},
\code{\link{point_plot}()},
\code{\link{stick_plot}()}
}
\concept{NEMO-MEDUSA plots}
