% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions.R
\name{type_in_month}
\alias{type_in_month}
\title{Condense Daily netcdf Files into a Month by Type}
\usage{
type_in_month(data, ...)
}
\arguments{
\item{data}{A dataframe containing the metadata of multiple netcdf files which share a type.}

\item{...}{Additional arguments passed to the relevant `get_*` function.}
}
\value{
The function returns a dataframe containing the monthly average shalllow and deep spatial grids for
variables of interest.
}
\description{
This function takes the metadata for multiple netcdf files and creates a spatial summary for the month.
\strong{For one file type only}.
}
\details{
The function bridges the step between extracting data from a netcdf file, and creating an average dataframe
for a whole month of NEMO-MEDUSA model outputs.

Different file types require a different get function. This function takes a collection of netcdf files of
the same time from the same month, and passes them to the correct `get_*` for data extraction. The results
are passed back to this function, before the estimates from different days at the same lat-lon-depth
combination are averaged to get a single number for the month.

Creating intermediate monthly objects allows the terabytes of data to be reduced without encountering memory issues.
Also, working on independent monthly packets of data means we can parallelise any data processing for speed.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_grid_T_}()},
\code{\link{get_grid_U_}()},
\code{\link{get_grid_V_}()},
\code{\link{get_grid_W_}()},
\code{\link{get_icemod_}()},
\code{\link{get_ptrc_T_}()},
\code{\link{get_rivers}()},
\code{\link{get_spatial}()},
\code{\link{interp_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
