% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{vectors_2_direction}
\alias{vectors_2_direction}
\title{Convert a U-V velocity field to speed and direction in degrees}
\usage{
vectors_2_direction(u, v)
}
\arguments{
\item{u}{A vector of Zonal currents (from West to East).}

\item{v}{A vector of Meridional currents (from South to North).}
}
\value{
a dataframe of two columns is returned. Speed contains the composite speed of both velocities on the same scale.
Direction is the resolved direction of the flow in degrees, 0 heads north, 90 East, 180 South, 270 West.
}
\description{
This function takes a vector of u and v velocities and calculates the direction and speed of the combined movement.
}
\details{
This function was lifted from the `Rsenal` package, where it was originally used to calculate wind speeds. All I've done
is built a wrapper which accounts for different conventions when describing wind and flow directions.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{Window}()},
\code{\link{calculate_depth_share}()},
\code{\link{empty}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights}()},
\code{\link{reproj}()},
\code{\link{sfc_as_cols}()},
\code{\link{stratify}()},
\code{\link{voronoi_grid}()}
}
\concept{NEMO-MEDUSA spatial tools}
