% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions.R
\name{whole_month}
\alias{whole_month}
\title{Condense Daily netcdf Files into a Monthly Summary}
\usage{
whole_month(data, crop, grid, ...)
}
\arguments{
\item{data}{A dataframe containing the metadata of multiple netcdf files from a common month.}

\item{...}{Additional arguments to be passed to get_* functions.}
}
\value{
The function returns a dataframe containing the monthly average shalllow and deep spatial grids for
\strong{all} the variables of interest in NEMO-MEDUSA outputs.
}
\description{
This function takes the metadata for multiple netcdf files. It then creates a common spatial summary for all
the variables of interest for a month.
}
\details{
The function takes the metadata for a collection of files which contain data from the same month. The files
are split into data packets which share the same file type, before being passed to `type_in_month` to be summarised.
`type_in_month` reduces the NEMO-MEDUSA model outputs from large arrays to effectively two matrices. The summaries for
each file type are returned to this function and get bound into a single dataframe. Points outside the project window are
removed before saving the dataframe for the month in "./Objects/Months/".

Creating intermediate monthly objects allows the terabytes of data to be reduced without encountering memory issues.
Also, working on independent monthly packets of data means we can parallelise any data processing for speed.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_grid_T_}()},
\code{\link{get_grid_U_}()},
\code{\link{get_grid_V_}()},
\code{\link{get_grid_W_}()},
\code{\link{get_icemod_}()},
\code{\link{get_ptrc_T_}()},
\code{\link{get_rivers}()},
\code{\link{get_spatial}()},
\code{\link{interp_month}()},
\code{\link{type_in_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
