% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{Window}
\alias{Window}
\title{Get Indices to Use When Clipping netcdf Files at Import}
\usage{
Window(file, w, e, s, n)
}
\arguments{
\item{file}{The full name of a netcdf file containing a longitude and latitude dimension.}

\item{w}{Degrees West to read from.}

\item{e}{Degrees East to read to.}

\item{s}{Degrees South to read from.}

\item{n}{Degrees North to read to.}
}
\value{
A list of three elements:
\itemize{
 \item{\emph{Lats -}}{ A vector of latitudes from `s` to `n`.}
 \item{\emph{Lons -}}{ A vector of longitudes from `w` to `e`.}
 \item{\emph{Limits -}}{ A dataframe containing the index to start reading from (Lon_start, Lat_start)
 and the length of the vector to read (Lon_count, Lat_count.}
 }
}
\description{
This function works out how much of a netcdf file to read, to capture the data between a given Lat-Lon window.
}
\details{
The function reads in a vector for both latitudes and longitudes, and tests whether each entry is within the specified
window. The max and min position in these vectors where the condition == TRUE are taken to define the ends of the window
to import. The vectors of latitudes and longitudes between these limits are kept, so they can be added to the variables
of interest during extraction.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{calculate_depth_share}()},
\code{\link{empty}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights}()},
\code{\link{reproj}()},
\code{\link{sfc_as_cols}()},
\code{\link{stratify}()},
\code{\link{vectors_2_direction}()},
\code{\link{voronoi_grid}()}
}
\concept{NEMO-MEDUSA spatial tools}
