% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{xyindex_to_nindex}
\alias{xyindex_to_nindex}
\title{Convert XY coordinates to a value index for a matrix}
\usage{
xyindex_to_nindex(x, y, nrow)
}
\arguments{
\item{x}{vector of row numbers for target values.}

\item{y}{vector of column numbers for target values.}

\item{nrow}{The number of rows in the matrix}
}
\value{
A vector of indices for values in a matrix.
}
\description{
This function converts a double index for a matrix (xy) into a single index for values. This allows vectorised subsetting 
of incomplete rows and columns.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{Window}()},
\code{\link{calculate_depth_share}()},
\code{\link{empty}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights}()},
\code{\link{reproj}()},
\code{\link{sfc_as_cols}()},
\code{\link{stratify}()},
\code{\link{vectors_2_direction}()},
\code{\link{voronoi_grid}()}
}
\concept{NEMO-MEDUSA spatial tools}
