% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{empty}
\alias{empty}
\title{Check Whether a Vector Contains Data}
\usage{
empty(x)
}
\arguments{
\item{x}{A vector of values to check.}
}
\value{
TRUE (if only NAs) or FALSE (if any entry is not NA).
}
\description{
This function provides a check for whether a vector of locations contains any data. This is useful
for checking whether a lat/lon pixel doesn't report any data for the deep zone. This is expected if the
sea floor at those coordinates is shallower than the maximum depth we set for the shallow zone.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{calculate_depth_share}()},
\code{\link{calculate_proximity_weight}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights}()},
\code{\link{netcdf_scheme_helpers}},
\code{\link{scheme_interp_slice}()},
\code{\link{scheme_strathE2E}()},
\code{\link{stratify}()},
\code{\link{voronoi_grid}()},
\code{\link{xyindex_to_nindex}()}
}
\concept{NEMO-MEDUSA spatial tools}
