% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions slabR.R
\name{extractors_slabR}
\alias{extractors_slabR}
\alias{get_grid_T_slabR}
\alias{get_ptrc_T_slabR}
\alias{get_grid_W_slabR}
\alias{get_grid_V_slabR}
\alias{get_grid_U_slabR}
\alias{get_icemod_slabR}
\title{Extract summaries from NEMO-MEDUSA arrays using slabR}
\usage{
get_grid_T_slabR(path, file, scheme, ice_scheme, space, ...)

get_ptrc_T_slabR(path, file, scheme, space, ...)

get_grid_W_slabR(path, file, scheme_w, space, ...)

get_grid_V_slabR(path, file, scheme, space, ...)

get_grid_U_slabR(path, file, scheme, space, ...)

get_icemod_slabR(path, file, scheme, ice_scheme, space, ...)
}
\arguments{
\item{path}{the path to the NEMO-MEDUSA model outputs.}

\item{file}{the name of a netcdf file containing the title variables.}

\item{scheme}{a summary scheme as expected by \code{array_w_mean()}.}

\item{ice_scheme}{a separate summary scheme for the 2D ice variables.}

\item{space}{a list object containing stuff....}

\item{...}{soaks up unused function arguments passed by the wrapper functions handling file architecture.}

\item{scheme_w}{a summary scheme as expected by \code{array_w_mean()}.}
}
\value{
A matrix with a column of group averages per title variable for a single day.
}
\description{
These functions read in target variables from NEMO-MEDUSA model outputs and return weighted averages
according to a summary scheme.
}
\details{
Each variable of interest in a netcdf file is imported, only reading within an x/y window specified
with \code{start3D} and \code{count3D}. The values are then passed to \code{array_w_mean()} to summarise according to a scheme.

Each model output file type contains a different set of variables, extracted by the relevant function variant:\tabular{ll}{
   File type         \tab Variables \cr
   grid_T_ \tab Salinity, temperature, sea ice concentration. \cr
   grid_U_ \tab Zonal currents. \cr
   grid_V_ \tab Meridional currents. \cr
   grid_W_ \tab Vertical velocity, vertical eddy diffusivitiy. \cr
   icemod_ \tab Ice presence, ice thickness, snow thickness. \cr
   ptrc_T_ \tab DIN, phytoplankton nitrogen content. \cr
}


Some function variants have different arguments:

grid_T_ and icemod_ functions accept an \code{ice_scheme}. Cryosphere data is contained in matrices not arrays.

grid_W_ expects it's own \code{scheme_w} as depth levels are different between these and other files.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{NEMO_MEDUSA}()},
\code{\link{extractors_1D}},
\code{\link{get_grid_T_StrathE2E}()},
\code{\link{get_grid_U_StrathE2E}()},
\code{\link{get_grid_V_StrathE2E}()},
\code{\link{get_grid_W_StrathE2E}()},
\code{\link{get_icemod_StrathE2E}()},
\code{\link{get_ptrc_T_StrathE2E}()},
\code{\link{get_spatial}()}
}
\concept{NEMO-MEDUSA variable extractors}
