% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions STRATHE2E.R
\name{get_rivers}
\alias{get_rivers}
\title{Get Rivers}
\usage{
get_rivers(File, Year, domain)
}
\arguments{
\item{File}{The name (with path) of a netcdf file containing river runoff data.}

\item{Year}{The year the file contains data for.}

\item{domain}{The polygon used to capture point estimates relevant to the model domain.}
}
\value{
A dataframe containing the total river runoff into the model domain by month with a year column.
}
\description{
This function reads river NEMO-MEDUSA river runoff files and reshapes for StrathE2E.
}
\details{
The river runoff file for a single year is imported. Points are checked to see if they fall in the model domain.
The points which do are summed by months.

NOTE! The river run off files provided to MiMeMo have estimates in open ocean. Runoff is channeled to cells in
the vicinity of river mouths so that single grid cells don't get the full amount. Of course river run off really
comes from the coast, so you may need to modify your domain polygon.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_grid_T_1D}()},
\code{\link{get_grid_T_StrathE2E}()},
\code{\link{get_grid_U_1D}()},
\code{\link{get_grid_U_StrathE2E}()},
\code{\link{get_grid_V_1D}()},
\code{\link{get_grid_V_StrathE2E}()},
\code{\link{get_grid_W_1D}()},
\code{\link{get_grid_W_StrathE2E}()},
\code{\link{get_icemod_StrathE2E}()},
\code{\link{get_ptrc_T_1D}()},
\code{\link{get_ptrc_T_StrathE2E}()},
\code{\link{get_spatial}()},
\code{\link{interp_month}()},
\code{\link{type_in_month_old}()},
\code{\link{type_in_month}()},
\code{\link{whole_month_old}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
