% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{get_spatial}
\alias{get_spatial}
\title{Get Latitudes, Longitudes, & Depths}
\usage{
get_spatial(file)

get_spatial(file)
}
\arguments{
\item{file}{The full name of a netcdf file.}
}
\value{
A list of three elements:
\itemize{
 \item{\emph{nc_lat -}}{ A matrix of latitudes which maps onto the first and second dimension of a NEMO-MEDUSA array.}
 \item{\emph{nc_lon -}}{ A matrix of longitudes which maps onto the first and second dimension of a NEMO-MEDUSA array.}
 \item{\emph{nc_depth -}}{ A vector of depths which match the third dimension of a NEMO-MEDUSA array.}
 }

A list of three elements:
\itemize{
 \item{\emph{nc_lat -}}{ A matrix of latitudes which maps onto the first and second dimension of a NEMO-MEDUSA array.}
 \item{\emph{nc_lon -}}{ A matrix of longitudes which maps onto the first and second dimension of a NEMO-MEDUSA array.}
 \item{\emph{nc_depth -}}{ A vector of depths which match the third dimension of a NEMO-MEDUSA array.}
 }
}
\description{
This function gets the latitudes, longitudes, and depths which define the spatial location of points in an array of NEMO-MEDUSA outputs.

This function gets the latitudes, longitudes, and depths which define the spatial location of points in an array of NEMO-MEDUSA outputs.
}
\details{
Each variable of interest in the netcdf file is imported, and then collected into a list.

Each variable of interest in the netcdf file is imported, and then collected into a list.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{detritus_month}()},
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_bio}()},
\code{\link{get_detritus}()},
\code{\link{get_ice}()},
\code{\link{get_merid}()},
\code{\link{get_rivers}()},
\code{\link{get_sea}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{interp_month}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}

Other NEMO-MEDUSA variable extractors: 
\code{\link{detritus_month}()},
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_bio}()},
\code{\link{get_detritus}()},
\code{\link{get_ice}()},
\code{\link{get_merid}()},
\code{\link{get_rivers}()},
\code{\link{get_sea}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{interp_month}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
