% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{NEMO_MEDUSA}
\alias{NEMO_MEDUSA}
\title{Condense NEMO MEDUSA Model Outputs into a Summary file}
\usage{
NEMO_MEDUSA(
  data,
  analysis,
  out_dir = "./Objects/Months",
  crop = NULL,
  grid = NULL,
  ...
)
}
\arguments{
\item{data}{A dataframe containing the metadata of multiple netcdf files from a common month.}

\item{...}{Additional arguments to be passed to get_* functions.}
}
\value{
The function returns a dataframe containing the monthly average shalllow and deep spatial grids for
\strong{all} the variables of interest in NEMO-MEDUSA outputs.
}
\description{
This function takes the metadata for multiple netcdf files. It then creates a single summary file for all
variables according to an analysis argument.
}
\details{
The function takes the metadata for a collection of files which contain data from the same month. The files
are split into data packets which share the same file type, before being passed to `temporal_operation` to be summarised.
`temporal_operation` reduces the NEMO-MEDUSA model outputs from large arrays to summaries depending on the analysis chosen.
The summaries for each file type are returned to this function and get bound into a single dataframe. 

StrathE2E analysis returns effectively two matrices. Points outside the project window are
removed before saving the dataframe for the month in "./Objects/Months/".

slabR uses a C++ routine to summarise arrays according to indices and a grouping scheme 
before saving the dataframe for the month in "./Objects/Months/".

1D analysis returns a vector of estimates at one pixel across all depths.

Creating intermediate monthly objects allows the terabytes of data to be reduced without encountering memory issues.
Also, working on independent monthly packets of data means we can parallelise any data processing for speed.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{extractors_1D}},
\code{\link{extractors_slabR}},
\code{\link{get_grid_T_StrathE2E}()},
\code{\link{get_grid_U_StrathE2E}()},
\code{\link{get_grid_V_StrathE2E}()},
\code{\link{get_grid_W_StrathE2E}()},
\code{\link{get_icemod_StrathE2E}()},
\code{\link{get_ptrc_T_StrathE2E}()},
\code{\link{get_spatial}()}
}
\concept{NEMO-MEDUSA variable extractors}
