% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{voronoi_grid}
\alias{voronoi_grid}
\title{Calculate the Domain Area per Grid Point}
\usage{
voronoi_grid(points, area)
}
\arguments{
\item{points}{A Simple Feature object og the grid points within the model domain.}

\item{area}{A Simple Feature object containing the model domain.}
}
\value{
the `points` object is returned, but instead of points, the geometry column now contains polygons representing the area closest to each point. A column for the size of this area is also gained.
}
\description{
This function takes an array of a variable, and an array of water thicknesses to perform a weighted average across depth. The depth
window to be averaged can be specified, so this function can be used to create both shallow and deep layers (or more for that matter).
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{Window}()},
\code{\link{calculate_depth_share}()},
\code{\link{empty}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights.old}()},
\code{\link{get_weights}()},
\code{\link{reproj}()},
\code{\link{sfc_as_cols}()},
\code{\link{stratify}()},
\code{\link{vectors_2_direction}()}
}
\concept{NEMO-MEDUSA spatial tools}
